Moved single node specific entities to nested module.
