+++ src/server/glommio_impl/processor/single_node.rs
@@ -1,11 +1,17 @@
 use super::{
     super::{
-        channel::Sender,
-        partition::{Partition, PartitionCreator, PartitionId, PartitionRequest},
-        worker::{ProcessorRequest, Task, TaskError, TaskResult, WorkerRequest},
-        Response,
+        super::{
+            channel::Sender,
+            partition::{single_node::PartitionRequest, Partition, PartitionCreator, PartitionId},
+            single_node::{Request, Response},
+            worker::{
+                single_node::{ProcessorRequest, WorkerRequest},
+                Task, TaskError, TaskResult,
+            },
+        },
+        worker::ResponseSender,
     },
-    worker::ResponseSender,
+    partition_consumer,
 };
 use glommio::{sync::RwLock, TaskQueueHandle};
 use std::{borrow::Cow, collections::HashMap, rc::Rc};
@@ -42,7 +48,7 @@ async fn handle_idempotent_requests<
     partitions: Rc<RwLock<HashMap<PartitionId, Rc<RwLock<P>>>>>,
     partition_id: PartitionId,
     request: PartitionRequest,
-) -> TaskResult<P> {
+) -> TaskResult<Response, P> {
     partitions
         .read()
         .await
@@ -61,7 +67,7 @@ async fn handle_requests<P: Partition<Request = PartitionRequest, Response = Res
     partitions: Rc<RwLock<HashMap<PartitionId, Rc<RwLock<P>>>>>,
     partition_id: PartitionId,
     request: PartitionRequest,
-) -> TaskResult<P> {
+) -> TaskResult<Response, P> {
     partitions
         .read()
         .await
@@ -83,7 +89,7 @@ async fn handle_create_partition<
     partitions: Rc<RwLock<HashMap<PartitionId, Rc<RwLock<P>>>>>,
     partition_id: PartitionId,
     partition_creator: PC,
-) -> TaskResult<P> {
+) -> TaskResult<Response, P> {
     if partitions
         .read()
         .await
@@ -106,203 +112,14 @@ async fn handle_create_partition<
     Ok(Response::PartitionCreated)
 }
 
-mod partition_consumer {
-    use super::super::super::{
-        partition::{Partition, PartitionCreator, PartitionId},
-        worker::{TaskError, TaskResult},
-        Response,
-    };
-    use glommio::sync::RwLock;
-    use std::{rc::Rc, time::Duration};
-
-    pub(crate) enum PartitionRemainder<P: Partition> {
-        Rc(Rc<RwLock<P>>),
-        RwLock(RwLock<P>),
-        Partition(P),
-        PartitionConsumed,
-    }
-
-    pub(crate) enum ConsumeMethod {
-        Close,
-        Remove,
-    }
-
-    pub(crate) struct PartitionConsumer<P, PC>
-    where
-        P: Partition,
-        PC: PartitionCreator<P>,
-    {
-        partition_remainder: Option<PartitionRemainder<P>>,
-        partition_id: PartitionId,
-        partition_creator: PC,
-
-        consume_method: ConsumeMethod,
-
-        retries: u32,
-        wait_duration: Duration,
-    }
-
-    impl<P, PC> PartitionConsumer<P, PC>
-    where
-        P: Partition,
-        PC: PartitionCreator<P>,
-    {
-        pub(crate) fn with_retries_and_wait_duration(
-            partition_remainder: PartitionRemainder<P>,
-            partition_id: PartitionId,
-            partition_creator: PC,
-            consume_method: ConsumeMethod,
-            retries: u32,
-            wait_duration: Duration,
-        ) -> Self {
-            Self {
-                partition_remainder: Some(partition_remainder),
-                partition_id,
-                partition_creator,
-                consume_method,
-                retries,
-                wait_duration,
-            }
-        }
-
-        pub(crate) fn new(
-            partition_remainder: PartitionRemainder<P>,
-            partition_id: PartitionId,
-            partition_creator: PC,
-            consume_method: ConsumeMethod,
-        ) -> Self {
-            Self::with_retries_and_wait_duration(
-                partition_remainder,
-                partition_id,
-                partition_creator,
-                consume_method,
-                5,
-                Duration::from_millis(100),
-            )
-        }
-
-        async fn remove_remainder(
-            &self,
-            partition_remainder: PartitionRemainder<P>,
-        ) -> Result<Option<PartitionRemainder<P>>, PartitionRemainder<P>> {
-            match partition_remainder {
-                PartitionRemainder::Rc(rc) => Rc::try_unwrap(rc)
-                    .map(|x| Some(PartitionRemainder::RwLock(x)))
-                    .map_err(PartitionRemainder::Rc),
-                PartitionRemainder::RwLock(rwlock) => rwlock
-                    .into_inner()
-                    .map(|x| Some(PartitionRemainder::Partition(x)))
-                    .map_err(|_| PartitionRemainder::PartitionConsumed),
-                PartitionRemainder::Partition(partition) => match &self.consume_method {
-                    ConsumeMethod::Remove => partition.remove(),
-                    ConsumeMethod::Close => partition.close(),
-                }
-                .await
-                .map(|_| None)
-                .map_err(|_| PartitionRemainder::PartitionConsumed),
-                PartitionRemainder::PartitionConsumed => {
-                    if let Ok(partition) = self
-                        .partition_creator
-                        .new_partition(&self.partition_id)
-                        .await
-                    {
-                        Ok(Some(PartitionRemainder::Partition(partition)))
-                    } else {
-                        Err(PartitionRemainder::PartitionConsumed)
-                    }
-                }
-            }
-        }
-
-        pub(crate) async fn consume(mut self) -> TaskResult<P> {
-            // Initial value of self.partition_remainder must be a Some
-            let mut last_iter_ok = false;
-
-            while let Some(remainder) = self.partition_remainder.take() {
-                self.partition_remainder = match self.remove_remainder(remainder).await {
-                    Err(partition_remainder) => {
-                        last_iter_ok = false;
-
-                        if self.retries.checked_sub(1).is_none() {
-                            None
-                        } else {
-                            glommio::timer::sleep(self.wait_duration).await;
-                            self.wait_duration *= 2;
-                            Some(partition_remainder)
-                        }
-                    }
-                    Ok(val) => {
-                        last_iter_ok = true;
-                        val
-                    }
-                }
-            }
-
-            if last_iter_ok {
-                Ok(Response::PartitionRemoved)
-            } else {
-                Err(TaskError::PartitionLost(self.partition_id))
-            }
-        }
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::{
-            super::super::super::partition::in_memory::{
-                Partition as InMemPartition, PartitionCreator as InMemPartitionCreator,
-            },
-            *,
-        };
-        use glommio::{LocalExecutorBuilder, Placement};
-        use std::collections::HashMap;
-
-        #[test]
-        fn test_partition_consumer() {
-            LocalExecutorBuilder::new(Placement::Unbound)
-                .spawn(|| async move {
-                    let partition_container = Rc::new(RwLock::new(HashMap::<
-                        PartitionId,
-                        Rc<RwLock<InMemPartition>>,
-                    >::new()));
-
-                    let partition_id = PartitionId {
-                        topic: "some_topic".into(),
-                        partition_number: 0,
-                    };
-
-                    partition_container.write().await.unwrap().insert(
-                        partition_id.clone(),
-                        Rc::new(RwLock::new(InMemPartition::new())),
-                    );
-
-                    let partition = partition_container
-                        .write()
-                        .await
-                        .unwrap()
-                        .remove(&partition_id)
-                        .unwrap();
-
-                    PartitionConsumer::new(
-                        PartitionRemainder::Rc(partition),
-                        partition_id,
-                        InMemPartitionCreator,
-                        ConsumeMethod::Remove,
-                    )
-                    .consume()
-                    .await
-                    .unwrap();
-                })
-                .unwrap();
-        }
-    }
-}
-
-async fn handle_remove_partition<P: Partition, PC: PartitionCreator<P>>(
+async fn handle_remove_partition<
+    P: Partition<Request = PartitionRequest, Response = Response>,
+    PC: PartitionCreator<P>,
+>(
     partitions: Rc<RwLock<HashMap<PartitionId, Rc<RwLock<P>>>>>,
     partition_id: PartitionId,
     partition_creator: PC,
-) -> TaskResult<P> {
+) -> TaskResult<Response, P> {
     let partition = partitions
         .write()
         .await
@@ -318,11 +135,12 @@ async fn handle_remove_partition<P: Partition, PC: PartitionCreator<P>>(
     )
     .consume()
     .await
+    .map(|_| Response::PartitionRemoved)
 }
 
-async fn partition_hierachy<P: Partition>(
+async fn partition_hierachy<P: Partition<Request = PartitionRequest, Response = Response>>(
     partitions: Rc<RwLock<HashMap<PartitionId, Rc<RwLock<P>>>>>,
-) -> TaskResult<P> {
+) -> TaskResult<Response, P> {
     let mut topic_to_partitions = HashMap::<Cow<'static, str>, Vec<u64>>::new();
 
     for key in partitions
@@ -341,12 +159,14 @@ async fn partition_hierachy<P: Partition>(
     Ok(Response::PartitionHierachy(topic_to_partitions))
 }
 
-impl<P, PC> crate::server::worker::Processor<P, ResponseSender<P>> for Processor<P, PC>
+impl<P, PC>
+    super::super::super::worker::Processor<P, Request, Response, ResponseSender<Response, P>>
+    for Processor<P, PC>
 where
     P: Partition<Request = PartitionRequest, Response = Response> + 'static,
     PC: PartitionCreator<P> + Clone + 'static,
 {
-    fn process(&self, task: Task<P, ResponseSender<P>>) {
+    fn process(&self, task: Task<P, Request, Response, ResponseSender<Response, P>>) {
         let (partitions, partition_creator) =
             (self.partitions.clone(), self.partition_creator.clone());
 
@@ -357,6 +177,7 @@ where
                         PartitionRequest::Read { offset: _ }
                         | PartitionRequest::LowestOffset
                         | PartitionRequest::HighestOffset => {
+                            let request = request;
                             handle_idempotent_requests(partitions, partition, request).await
                         }
 
@@ -432,18 +253,36 @@ where
 
 #[cfg(test)]
 mod tests {
-    use super::super::worker::new_task;
-    use super::Processor;
-    use crate::server::{
-        channel::Receiver,
-        partition::{
-            in_memory::{Partition, PartitionCreator},
-            PartitionId, PartitionRequest,
+    use super::{
+        super::super::{
+            super::super::server::{
+                channel::Receiver,
+                glommio_impl::worker::{ResponseReceiver, ResponseSender},
+                partition::{
+                    in_memory::{Partition, PartitionCreator},
+                    single_node::PartitionRequest,
+                    PartitionId,
+                },
+                single_node::{Request, Response},
+                worker::{
+                    single_node::{ProcessorRequest, WorkerRequest},
+                    Processor as BaseProcessor, Task, TaskError,
+                },
+            },
+            worker::new_task,
         },
-        worker::{Processor as BaseProcessor, ProcessorRequest, TaskError, WorkerRequest},
-        Response,
+        Processor,
     };
     use glommio::{executor, Latency, LocalExecutorBuilder, Placement, Shares};
+
+    fn new_single_node_task(
+        worker_request: WorkerRequest,
+    ) -> (
+        Task<Partition, Request, Response, ResponseSender<Response, Partition>>,
+        ResponseReceiver<Response, Partition>,
+    ) {
+        new_task::<Partition, Request, Response>(worker_request.into())
+    }
 
     #[test]
     fn test_processor() {
@@ -466,7 +305,7 @@ mod tests {
                     partition_number: 2,
                 };
 
-                let (lowest_offset_task, recv) = new_task::<Partition>(WorkerRequest::Partition {
+                let (lowest_offset_task, recv) = new_single_node_task(WorkerRequest::Partition {
                     partition: partition_id_1.clone(),
                     request: PartitionRequest::LowestOffset,
                 });
@@ -483,12 +322,12 @@ mod tests {
                 }
 
                 let (create_partition_task_1, recv_1) =
-                    new_task::<Partition>(WorkerRequest::Processor(
+                    new_single_node_task(WorkerRequest::Processor(
                         ProcessorRequest::CreatePartition(partition_id_1.clone()),
                     ));
 
                 let (create_partition_task_2, recv_2) =
-                    new_task::<Partition>(WorkerRequest::Processor(
+                    new_single_node_task(WorkerRequest::Processor(
                         ProcessorRequest::CreatePartition(partition_id_2.clone()),
                     ));
 
@@ -503,7 +342,7 @@ mod tests {
                 let sample_record_bytes: &[u8] = b"Lorem ipsum dolor sit amet.";
 
                 let (append_record_task_1, recv_1) =
-                    new_task::<Partition>(WorkerRequest::Partition {
+                    new_single_node_task(WorkerRequest::Partition {
                         partition: partition_id_1.clone(),
                         request: PartitionRequest::Append {
                             record_bytes: sample_record_bytes.into(),
@@ -519,13 +358,13 @@ mod tests {
                 }
 
                 let (highest_offset_task_1, recv_1) =
-                    new_task::<Partition>(WorkerRequest::Partition {
+                    new_single_node_task(WorkerRequest::Partition {
                         partition: partition_id_1.clone(),
                         request: PartitionRequest::HighestOffset,
                     });
 
                 let (highest_offset_task_2, recv_2) =
-                    new_task::<Partition>(WorkerRequest::Partition {
+                    new_single_node_task(WorkerRequest::Partition {
                         partition: partition_id_2.clone(),
                         request: PartitionRequest::HighestOffset,
                     });
@@ -545,11 +384,10 @@ mod tests {
                     assert!(false, "Wrong response type for HighestOffset request");
                 }
 
-                let (read_record_task_1, recv_1) =
-                    new_task::<Partition>(WorkerRequest::Partition {
-                        partition: partition_id_1.clone(),
-                        request: PartitionRequest::Read { offset: 0 },
-                    });
+                let (read_record_task_1, recv_1) = new_single_node_task(WorkerRequest::Partition {
+                    partition: partition_id_1.clone(),
+                    request: PartitionRequest::Read { offset: 0 },
+                });
 
                 processor.process(read_record_task_1);
 
@@ -563,7 +401,7 @@ mod tests {
                     assert!(false, "Wrong response type for Read request");
                 }
 
-                let (partition_hierachy_task, recv) = new_task::<Partition>(
+                let (partition_hierachy_task, recv) = new_single_node_task(
                     WorkerRequest::Processor(ProcessorRequest::PartitionHierarchy),
                 );
 
@@ -579,12 +417,12 @@ mod tests {
                 }
 
                 let (remove_partition_task_1, recv_1) =
-                    new_task::<Partition>(WorkerRequest::Processor(
+                    new_single_node_task(WorkerRequest::Processor(
                         ProcessorRequest::RemovePartition(partition_id_1.clone()),
                     ));
 
                 let (remove_partition_task_2, recv_2) =
-                    new_task::<Partition>(WorkerRequest::Processor(
+                    new_single_node_task(WorkerRequest::Processor(
                         ProcessorRequest::RemovePartition(partition_id_2.clone()),
                     ));
                 processor.process(remove_partition_task_1);
@@ -595,12 +433,12 @@ mod tests {
 
                 // test drop
                 let (create_partition_task_1, recv_1) =
-                    new_task::<Partition>(WorkerRequest::Processor(
+                    new_single_node_task(WorkerRequest::Processor(
                         ProcessorRequest::CreatePartition(partition_id_1.clone()),
                     ));
 
                 let (create_partition_task_2, recv_2) =
-                    new_task::<Partition>(WorkerRequest::Processor(
+                    new_single_node_task(WorkerRequest::Processor(
                         ProcessorRequest::CreatePartition(partition_id_2.clone()),
                     ));
